fixies

+++ CMP102_Porj_code_Spring2020/Restaurant/Rest/Restaurant.h
@@ -66,7 +66,7 @@ private:
 	
 	int autoPromotedCount;
 	int autoPromotionSteps;
-	int injuryProbability;
+	double injuryProbability;
 	int urgentSteps;
 	int restSteps;
